refactor: extract resolver value-coercion helpers to clear the size cap
