+++ data/feature_index.json
@@ -1210,6 +1210,6 @@
       ]
     }
   ],
-  "generated_at": "2026-06-09T23:34:58.498598Z",
+  "generated_at": "2026-06-10T06:56:26.275861Z",
   "schema_version": 1
 }
