+++ src/synthorg/settings/resolver.py
@@ -35,6 +35,12 @@ from synthorg.observability.events.settings import (
     SETTINGS_VALIDATION_FAILED,
 )
 from synthorg.settings._resolver_batch_reads import resolve_bridge_fields
+from synthorg.settings._resolver_coercions import (
+    _build_budget_alerts,
+    _coerce_batch_size,
+    _coerce_vram_gb,
+    _parse_bool,
+)
 from synthorg.settings.bridge_configs import (
     A2ABridgeConfig,
     ApiBridgeConfig,
@@ -59,7 +65,7 @@ if TYPE_CHECKING:
     # (mirrors, registry seeds, kill switches), so the bridge-read
     # signatures name them without importing them.
     from synthorg.api.config import ApiConfig
-    from synthorg.budget.config import BudgetAlertConfig, BudgetConfig
+    from synthorg.budget.config import BudgetConfig
     from synthorg.config.agent_schema import AgentConfig
     from synthorg.config.schema import ProviderConfig, RootConfig
     from synthorg.core.company_departments import Department
@@ -1178,120 +1184,3 @@ class ConfigResolver:
             error_backoff_seconds=values["dispatcher_error_backoff_seconds"],
             max_consecutive_errors=values["dispatcher_max_consecutive_errors"],
         )
-
-
-def _coerce_vram_gb(value: object) -> float:
-    """Coerce a parsed JSON value to a numeric VRAM threshold.
-
-    Plain ``float(value)`` would accept ``True`` / ``False`` (because
-    ``bool`` is an ``int`` subclass and ``int`` is float-coercible), so
-    a payload like ``[true, 64]`` would silently become ``(1.0, 64)``
-    and pass the remaining shape checks. Reject booleans and
-    non-numeric types at the boundary so invalid stored settings fail
-    deterministically.
-
-    Returns:
-        The value coerced to ``float``, guaranteed to be a non-boolean
-        numeric type.
-
-    Raises:
-        TypeError: If *value* is a ``bool`` or any non-numeric type.
-    """
-    if isinstance(value, bool):
-        msg = f"vram_gb must be numeric, got bool {value!r}"
-        raise TypeError(msg)
-    if not isinstance(value, int | float):
-        msg = f"vram_gb must be numeric, got {type(value).__name__} {value!r}"
-        raise TypeError(msg)
-    return float(value)
-
-
-def _coerce_batch_size(value: object) -> int:
-    """Coerce a parsed JSON value to an ``int`` batch size, rejecting bad shapes.
-
-    Plain ``int(value)`` would silently truncate ``64.9`` to ``64`` and
-    accept ``True`` / ``False`` (which are ``int`` subclasses), so a
-    typo in ``memory.fine_tune_vram_batch_table`` would apply with a
-    different value than the operator configured. Reject those at the
-    boundary so invalid stored settings fail deterministically.
-
-    Returns:
-        The value coerced to ``int``, guaranteed to be a whole-number
-        non-boolean numeric type.
-
-    Raises:
-        TypeError: If *value* is a ``bool`` or any non-numeric type.
-        ValueError: If *value* is a fractional (non-integer) float.
-    """
-    if isinstance(value, bool):
-        msg = f"batch_size must be an integer, got bool {value!r}"
-        raise TypeError(msg)
-    if isinstance(value, int):
-        return value
-    if isinstance(value, float):
-        if not value.is_integer():
-            msg = f"batch_size must be an integer, got fractional float {value!r}"
-            raise ValueError(msg)
-        return int(value)
-    msg = f"batch_size must be an integer, got {type(value).__name__} {value!r}"
-    raise TypeError(msg)
-
-
-def _build_budget_alerts(warn: int, crit: int, stop: int) -> BudgetAlertConfig:
-    """Construct ``BudgetAlertConfig`` with ordering validation.
-
-    Args:
-        warn: Warning threshold percent.
-        crit: Critical threshold percent.
-        stop: Hard-stop threshold percent.
-
-    Returns:
-        A validated ``BudgetAlertConfig``.
-
-    Raises:
-        ValueError: If the thresholds violate the ordering constraint
-            (``warn < crit < stop``).
-    """
-    from pydantic import ValidationError  # noqa: PLC0415
-
-    from synthorg.budget.config import BudgetAlertConfig  # noqa: PLC0415
-
-    try:
-        return BudgetAlertConfig(warn_at=warn, critical_at=crit, hard_stop_at=stop)
-    except ValidationError as exc:
-        logger.warning(
-            SETTINGS_VALIDATION_FAILED,
-            namespace="budget",
-            key="_alerts",
-            reason="threshold_ordering",
-        )
-        msg = "Budget alert thresholds must satisfy warn < critical < hard_stop"
-        raise ValueError(msg) from exc
-
-
-_BOOL_TRUE = frozenset({"true", "1"})
-_BOOL_FALSE = frozenset({"false", "0"})
-
-
-def _parse_bool(value: str) -> bool:
-    """Parse a string into a boolean.
-
-    Accepts ``"true"``/``"false"``/``"1"``/``"0"``
-    (case-insensitive).
-
-    Args:
-        value: String to parse.
-
-    Returns:
-        The parsed boolean.
-
-    Raises:
-        ValueError: If the string is not a recognised boolean.
-    """
-    lower = value.lower()
-    if lower in _BOOL_TRUE:
-        return True
-    if lower in _BOOL_FALSE:
-        return False
-    msg = "Value is not a recognized boolean string"
-    raise ValueError(msg)
